+++ samples/spring-webflux-security-hybrid-usage/src/main/java/sample/spring/webflux/hybrid/SecurityConfiguration.java
@@ -3,7 +3,7 @@
  * <p>
  * SPDX-License-Identifier: Apache-2.0
  */
-package sample.spring.webflux.xsuaa;
+package sample.spring.webflux.hybrid;
 
 import com.sap.cloud.security.spring.config.IdentityServiceConfiguration;
 import com.sap.cloud.security.spring.config.IdentityServicesPropertySourceFactory;
