+++ samples/spring-webflux-security-hybrid-usage/src/main/resources/application.yml
@@ -7,7 +7,7 @@ logging.level:
   org.springframework: ERROR          # set to DEBUG to see all beans loaded and auto-config conditions met.
   org.springframework.security: DEBUG # set to ERROR for production setups. 
   org.springframework.webflux: DEBUG      # set to ERROR for production setups.
-  sample.spring.webflux.xsuaa: INFO
+  sample.spring.webflux.hybrid: INFO
 spring:
   main:
     allow-bean-definition-overriding: true
