+++ samples/spring-webflux-security-hybrid-usage/src/test/java/sample/spring/webflux/hybrid/ApplicationTest.java
@@ -3,7 +3,7 @@
  * <p>
  * SPDX-License-Identifier: Apache-2.0
  */
-package sample.spring.webflux.xsuaa;
+package sample.spring.webflux.hybrid;
 
 import com.sap.cloud.security.spring.autoconfig.HybridIdentityServicesAutoConfiguration;
 import org.junit.Test;
